\name{epinet-internal}
\Rdversion{2.10}
\title{Internal epinet Objects}
\alias{itimestartvalues}
\alias{etimestartvalues}
\alias{CreateCompactMatrix}
\alias{removesusceptibles}
\alias{buildepifromoutput}
\alias{epi2newick.int}
\alias{plotepitree}
\description{
  Internal epinet functions.
}
\details{
  These are not intended to be called by the user and
  are only used by other functions.
}
%\seealso{}
\keyword{internal}
