\name{epinet-internal}
\Rdversion{2.10}
\title{Internal epinet Objects}
\alias{itimestartvalues}
\alias{etimestartvalues}
\description{
  Internal epinet functions.
}
\details{
  These are not intended to be called by the user and
  are only used by other functions.
}
%\seealso{}
\keyword{internal}
